refactor: move components to src/components
