+++ charms/kfp-ui/src/charm.py
@@ -27,7 +27,7 @@ from charms.kubeflow_dashboard.v0.kubeflow_dashboard_links import (
 from lightkube.resources.rbac_authorization_v1 import ClusterRole, ClusterRoleBinding
 from ops import BoundEvent, CharmBase, main
 
-from pebble_components import MlPipelineUiInputs, MlPipelineUiPebbleService
+from components.pebble_components import MlPipelineUiInputs, MlPipelineUiPebbleService
 
 logger = logging.getLogger(__name__)
